requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
openai>=1.12.0
GitPython>=3.1.40
requests>=2.31.0
pyyaml>=6.0.1
mcp>=0.9.0
